Flagging Initialisation Data Listener
On each construction of an ElectronStore instance, the initDataListener method is called. This causes duplicate ipcMain handlers for the 'electron-store-get-data' channel. Found that this causes an Electron EventEmitter warning of 'MaxListenersExceededWarning'. By setting an initialiser flag, events are handled as needed and does not cause this warning. Additionally if someone were to properly exceed this (in my case I was at 11), by dynamically creating larger amounts of stores (which could be possible in larger scale applications, although not recommended), this could have dangerous and odd side-effects.